+++ app.js
@@ -22,4 +22,5 @@ app.use((err, req, res, next) => {
   res.status(status).json({ message })
 })
 
+
 export default app;
