+++ models/contacts.js
@@ -2,7 +2,6 @@ import fs from "fs/promises";
 import path from "path";
 import { nanoid } from "nanoid";
 
-
 const contactsPath = path.resolve('models','contacts.json');
 const updateContactStorage = contacts => fs.writeFile(contactsPath, JSON.stringify(contacts, null, 2));
