+++ routes/api/contacts.js
@@ -1,5 +1,6 @@
 import express from "express";
 import contactsController from "../../controllers/contactsController.js"
+import dataValidate from "../../validators/contactsValidator.js";
 
 const router = express.Router();
 
@@ -7,10 +8,10 @@ router.get('/', contactsController.getAll);
 
 router.get('/:contactId', contactsController.getById);
 
-router.post('/', contactsController.add);
+router.post('/', dataValidate, contactsController.add);
 
 router.delete('/:contactId', contactsController.deleteById);
 
-router.put('/:contactId', contactsController.updateById);
+router.put('/:contactId', dataValidate, contactsController.updateById);
 
 export default router;
